chore: split current eth_calls collector into focused submodules
